// frontend/systeme-interne/src/app/models/ProfModel/departement.ts
export enum Departement {
  Département_de_Biologie = 'Département_de_Biologie',
  Département_de_Chimie = 'Département_de_Chimie',
  Département_de_Géologie = 'Département_de_Géologie',
  Département_d_Informatique = 'Département_d_Informatique',
  Département_de_Mathématiques = 'Département_de_Mathématiques',
  Département_de_Physique = 'Département_de_Physique',
}